Seperate samples into different files, simplify snippet.
